Testing AMD Module with requireJS
