+++ index.html
@@ -7,6 +7,7 @@
     <meta http-equiv="X-UA-Compatible" content="ie=edge">
     <title>Document</title>
     <link rel="stylesheet" href="src/styles/css/main.css">
+    <script src="./node_modules/requirejs/require.js"></script>
 </head>
 
 <body>
