+++ package.json
@@ -15,6 +15,7 @@
   "devDependencies": {
     "concurrently": "^3.6.1",
     "node-sass": "^4.9.3",
+    "requirejs": "^2.3.5",
     "typescript": "^3.0.1"
   }
 }
